requirements.txt
fastmcp>=2.0
openai>=1.0
python-dotenv
colorama
fastapi
uvicorn[standard]
websockets
aiosqlite
